CapstoneDesign_Server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
mediapipe
opencv-python-headless
openai
numpy
python-dotenv
openai-whisper
praat-parselmouth